+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning][semver].
 
 ## [Unreleased]
 
+### Changed
+
+- TMS-318: Selected materials bypass taxonomy selection #43
+- TMS-318: Indicate current filter for screen readers #43
+
 ## [1.6.1] - 2021-10-25
 
 - Update translations